+++ client/extensions/clipper/clipper.constants.ts
@@ -1,0 +1,1 @@
+export const csuiSelector = "plasmo-csui";
